+++ src/L10NSharp/Translators/TranslatorBase.cs
@@ -1,10 +1,10 @@
 using System;
 using L10NSharp.Utility;
 
-namespace L10NSharp.Windows.Forms.Translators
+namespace L10NSharp.Translators
 {
 	/// ----------------------------------------------------------------------------------------
-	internal abstract class TranslatorBase
+	public abstract class TranslatorBase
 	{
 		/// ------------------------------------------------------------------------------------
 		protected string m_srcCultureId;